requirements.txt
asgiref 
certifi==2021.5.30
cffi 
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django 
django-allauth==0.44.0
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-datetime-widget==0.9.3
django-filter==2.4.0
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-pagedown==2.2.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna 
markdown2==2.3.1
oauthlib==3.1.1
olefile==0.46
phonenumbers==8.12.26
phonenumberslite==8.12.26
Pillow 
psycopg2 
pycparser 
PyJWT 
pypng==0.0.20
PyQRCode==1.2.1
python-magic==0.4.24
python3-openid==3.2.0
pytz 
requests==2.25.1
requests-oauthlib==1.3.0
six 
sqlparse 
urllib3==1.26.5
wincertstore==0.2
